bugfix whiptail

+++ includes/functions.sh
@@ -899,17 +899,20 @@ function choose_media_folder_plexdrive() {
         "Medias à ajouter pour ${USER} (Barre espace pour la sélection)" 28 60 17 \
         $(cat /tmp/menumedia.txt) 3>&1 1>&2 2>&3
     )
-    touch "$MEDIASPERUSER"
-    for MEDDOCKER in $MEDIASTOINSTALL; do
-      echo -e "	${GREEN}* $(echo $MEDDOCKER | tr -d '"')${NC}"
-      echo $(echo ${MEDDOCKER} | tr -d '"') >>$MEDIASPERUSER
-    done
-    for line in $(cat $MEDIASPERUSER); do
-      line=$(echo ${line} | sed 's/\(.\)/\U\1/')
-      create_dir "${HOME}/local/${line}"
-      create_dir "/mnt/rclone/${USER}/${line}"
-    done
-    rm /tmp/menumedia.txt
+    exitstatus=$?
+    if [ $exitstatus = 0 ]; then
+      touch "$MEDIASPERUSER"
+      for MEDDOCKER in $MEDIASTOINSTALL; do
+        echo -e "	${GREEN}* $(echo $MEDDOCKER | tr -d '"')${NC}"
+        echo $(echo ${MEDDOCKER} | tr -d '"') >>$MEDIASPERUSER
+      done
+      for line in $(cat $MEDIASPERUSER); do
+        line=$(echo ${line} | sed 's/\(.\)/\U\1/')
+        create_dir "${HOME}/local/${line}"
+        create_dir "/mnt/rclone/${USER}/${line}"
+      done
+      rm /tmp/menumedia.txt
+    fi
   fi
   mkdir -p "${HOME}/filebot"
   echo ""
